pyproject.toml
[project]
name = "datatech-2025-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2.3",
    "pydantic>=2.0.0",
    "sf-hamilton[visualization]>=1.38.0",
    "prophet>=1.1.4",
    "matplotlib>=3.7.0",
    "fastprogress>=1.0.3",
    "ipynb>=0.5.1",
    "ipykernel>=6.29.5",
    "shap>=0.47.2",
    "xgboost>=3.0.1",
]
